+++ modules/sys/templates.js
@@ -51,6 +51,7 @@ export const preloadHandlebarsTemplates = () => {
         `${templatePath}/actor/parts/setup/setup-talents.hbs`,
         `${templatePath}/actor/parts/skills/base-skills.hbs`,
 
+        `${templatePath}/chat/alert.hbs`,
         `${templatePath}/chat/attack.hbs`,
         `${templatePath}/chat/check.hbs`,
         `${templatePath}/chat/crit.hbs`,
@@ -59,7 +60,6 @@ export const preloadHandlebarsTemplates = () => {
         `${templatePath}/chat/defend.hbs`,
         `${templatePath}/chat/initNote.hbs`,
         `${templatePath}/chat/fumble.hbs`,
-        `${templatePath}/chat/top.hbs`,
         `${templatePath}/chat/trauma.hbs`,
         `${templatePath}/chat/chat-save.hbs`,
         `${templatePath}/chat/skill.hbs`,