+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.4.20] - 2024-08-07
 ### Added
 - Reset Weapon Count shifts actor init to half weapon speed.
 - Tenacity alerts and auto-saves.